Extended test_powermodels_io()
and fixed bugs in powermodels_io()

+++ edisgo/flex_opt/check_tech_constraints.py
@@ -1071,7 +1071,7 @@ def _lv_allowed_voltage_limits(edisgo_obj, lv_grids=None, mode=None):
 
         # get all secondary sides and buses in grids
         buses_dict = {}
-        secondary_sides = pd.Series()
+        secondary_sides = pd.Series(dtype="float64")
         for grid in lv_grids:
             secondary_side = grid.station.index[0]
             if secondary_side in buses_in_pfa:
